requirements.txt
PySSTV
Pillow
flask
playsound
werkzeug
